+++ training/urls.py
@@ -22,11 +22,9 @@ from django.contrib.staticfiles.views import serve
 from django.urls import include, path
 from django.views.generic.base import RedirectView, TemplateView
 from rest_framework import routers
-from rest_framework.documentation import include_docs_urls
+from rest_framework.schemas import get_schema_view
 
-from myapp import views
-
-API_TITLE = 'Demo API'
+from myapp import views, schemas, __title__ as API_TITLE
 
 admin.autodiscover()
 
@@ -38,7 +36,8 @@ router.register(r'instructors', views.InstructorViewSet)
 
 urlpatterns = [
 
-    path('', RedirectView.as_view(url='/v1', permanent=False)),
+    # redirect the base URL to something useful, the Swagger UI:
+    path('', RedirectView.as_view(url='/swagger-ui/', permanent=False)),
     path('v1/', include(router.urls)),
     # TODO: Is there a Router than can generate these for me? If not, create one.
     # course relationships:
@@ -74,8 +73,16 @@ urlpatterns = [
     path('v1/instructors/<pk>/<related_field>/',
         views.InstructorViewSet.as_view({'get': 'retrieve_related'}), # a toMany relationship
         name='instructor-related'),
-    # swagger UI
-    path('v1/openapi/', TemplateView.as_view(template_name="index.html")),
+    # path /v1/openapi is the OAS 3.0 schema document for this app.
+    # path /swagger-ui is the swagger user interface that uses that schema document.
+    # (v1/openapi and swagger-ui fail if the user is not logged in as the routers are not public.
+    # so set public=True since the endpoint documentation is not a secret, just the content.)
+    path('v1/openapi', get_schema_view(generator_class=schemas.SchemaGenerator, public=True),
+         name='openapi-schema'),
+    path('swagger-ui/', TemplateView.as_view(
+        template_name='swagger-ui.html',
+        extra_context={'schema_url': 'openapi-schema', 'title': API_TITLE}
+    ), name='swagger-ui'),
     # The default request_uri is /oauth2-redirect.html (no /static prefix) so just pass it into staticfiles serve():
     path('oauth2-redirect.html', serve, {'path': 'oauth2-redirect.html'}),
     # browseable API and admin stuff. TODO: Consider leaving out except when debugging.